% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{apply_indent}
\alias{apply_indent}
\title{Apply indention to code lines}
\usage{
apply_indent(lines, arr, spaces, trim)
}
\arguments{
\item{lines}{array of code lines}

\item{arr}{an array returned by \code{\link{count_indent_lines}}}

\item{trim}{if this is true, left trimming is done}
}
\value{
an edited array of code lines
}
\description{
Apply indention to code lines
}
\seealso{
Other code formatting helper functions: 
\code{\link{count_indent_lines}()},
\code{\link{justify_line}()}
}
\concept{code formatting helper functions}
