% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{assert_one_of}
\alias{assert_one_of}
\title{Check if argument is one of the allowed options}
\usage{
assert_one_of(arg, allowed)
}
\arguments{
\item{arg}{the argument to check}

\item{allowed}{allowed options}
}
\value{
TRUE if check succeeds, throw error otherwise
}
\description{
Check if argument is one of the allowed options
}
