% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis.R
\name{check_split}
\alias{check_split}
\title{Check that code split matches original code}
\usage{
check_split(code, rec)
}
\arguments{
\item{code}{full code}

\item{rec}{split}

\item{verbose}{should the message be printed}
}
\value{
same as \code{readLines(file)}
}
\description{
Check that code split matches original code
}
