% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan_momentum_draws.R
\name{cmdstan_momentum_draws}
\alias{cmdstan_momentum_draws}
\title{Get momentum draws from a cmdstanr fit
object run with save_latent_dynamics = TRUE}
\usage{
cmdstan_momentum_draws(fit)
}
\arguments{
\item{fit}{A cmdstan fit object run with \code{save_latent_dynamics = TRUE}}
}
\value{
A posterior matrix of momentum draws
}
\description{
Get momentum draws from a cmdstanr fit
object run with save_latent_dynamics = TRUE
}
