% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan_unconstrained_draws.R
\name{cmdstan_unconstrained_draws}
\alias{cmdstan_unconstrained_draws}
\title{Get matrix of unconstrained draws from a cmdstanr fit
object run with save_latent_dynamics = TRUE}
\usage{
cmdstan_unconstrained_draws(fit)
}
\arguments{
\item{fit}{A cmdstan fit object run with save_latent_dynamics = TRUE}
}
\value{
A posterior matrix of unconstrained draws
}
\description{
Get matrix of unconstrained draws from a cmdstanr fit
object run with save_latent_dynamics = TRUE
}
