% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{count_indent_lines}
\alias{count_indent_lines}
\title{Count how much each line should be indented}
\usage{
count_indent_lines(lines)
}
\arguments{
\item{lines}{array of code lines}
}
\value{
an integer array with same length as \code{lines}
}
\description{
Count how much each line should be indented
}
\seealso{
Other code formatting helper functions: 
\code{\link{apply_indent}()},
\code{\link{justify_line}()}
}
\concept{code formatting helper functions}
