% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{create_model}
\alias{create_model}
\title{Create a model}
\usage{
create_model(stan_file, ...)
}
\arguments{
\item{stan_file}{path to a \code{.stan} file that describes the model}

\item{...}{compilation options}
}
\value{
a CmdStanModel object
}
\description{
Create a model
}
