% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{data_8schools}
\alias{data_8schools}
\title{The eight schools data set}
\description{
The eight schools data set
}
\references{
Rubin, D. B. (1981).
Estimation in Parallel randomized experiments.
Journal of Educational Statistics, 6, 377-401
}
\seealso{
Other Stan data and models: 
\code{\link{model_8schools}}
}
\concept{Stan data and models}
