% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{ensure_leading_linebreak}
\alias{ensure_leading_linebreak}
\title{Add empty line to beginning of string if it doesn't exist already}
\usage{
ensure_leading_linebreak(s)
}
\arguments{
\item{s}{a string}
}
\value{
modified string
}
\description{
Add empty line to beginning of string if it doesn't exist already
}
\seealso{
Other code formatting helper functions: 
\code{\link{format_code_helpers}}
}
\concept{code formatting helper functions}
