% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis_helper.R
\name{find_block}
\alias{find_block}
\title{Find a block from Stan code}
\usage{
find_block(code, block, check_valid = TRUE)
}
\arguments{
\item{code}{Stan code as a string}

\item{block}{name of block}

\item{check_valid}{Should the validity of the \code{block} name be checked?}
}
\value{
a named list with three parts
}
\description{
Find a block from Stan code
}
