% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_divergences.R
\name{fit_divergences}
\alias{fit_divergences}
\title{Analyze a cmdstanr fit to figure out where the divergences are coming from}
\usage{
fit_divergences(fit)
}
\arguments{
\item{fit}{A stan fit to analyze}
}
\value{
Another \code{cmdstanr} fit informing you about your divergences
}
\description{
Analyze a cmdstanr fit to figure out where the divergences are coming from
}
