% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{format_code}
\alias{format_code}
\title{Format Stan code}
\usage{
format_code(code, spaces = 2)
}
\arguments{
\item{code}{code as a string}

\item{spaces}{number of spaces to add to indenting after each opening
curly bracket (default = 2)}
}
\value{
updated code as a string
}
\description{
This formatting function
\itemize{
\item indents code based on opening and closing curly brackets
\item removes unnesessacy whitespace
}
}
\seealso{
Other code formatting functions: 
\code{\link{indent_code}()},
\code{\link{place_includes}()},
\code{\link{trim_code}()}
}
\concept{code formatting functions}
