% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{format_code_helpers}
\alias{format_code_helpers}
\alias{format_part}
\alias{justify_left}
\alias{justify_line}
\alias{locate_closing_bracket}
\title{Helper functions for code formatting}
\usage{
format_part(code, spaces, indent)

justify_left(code, indent)

justify_line(line, indent)

locate_closing_bracket(code)
}
\arguments{
\item{code}{code as a string}

\item{spaces}{number of spaces to add to indenting after each opening
curly bracket (default = 2)}

\item{indent}{base level of indentation}

\item{line}{line to format}
}
\value{
updated code as a string
}
\description{
\itemize{
\item \code{format_part} is a helper function used recursively by
\code{format_code}
\item \code{justlify_left} is the base case of the recursion, which
justifies the code left at a given number of indentation spaces
\item \code{justify_line} first trims a line by removing leading and
trailing whitespace, and then justifies it correctly
\item \code{locate_closing_bracket} finds the correct closing
bracket
}
}
\seealso{
Other code formatting helper functions: 
\code{\link{ensure_leading_linebreak}()}
}
\concept{code formatting helper functions}
