% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis.R
\name{full_stan_code}
\alias{full_stan_code}
\title{Get full Stan code with includes placed in}
\usage{
full_stan_code(code, ...)
}
\arguments{
\item{code}{Stan code to be formatted, or name of a file where the
code is.}

\item{...}{additional arguments to \code{\link{format_code}}}
}
\value{
program code as a formatted string
}
\description{
Get full Stan code with includes placed in
}
