% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis_helper.R
\name{get_block}
\alias{get_block}
\title{Return a block from Stan code}
\usage{
get_block(code, block)
}
\arguments{
\item{code}{Stan code as a string}

\item{block}{name of block}
}
\value{
a named list with three parts
}
\description{
Return a block from Stan code
}
