% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis.R
\name{get_functions}
\alias{get_functions}
\title{Get names of all user-defined functions}
\usage{
get_functions(code, ...)
}
\arguments{
\item{code}{full stan code}

\item{...}{additional arguments to \code{\link{place_includes}}}
}
\value{
a data frame
}
\description{
Get names of all user-defined functions
}
\seealso{
Other code analysis functions: 
\code{\link{get_stan_code}()}
}
\concept{code analysis functions}
