% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{get_stan_code}
\alias{get_stan_code}
\title{Get full Stan code from a .stan file, possibly containing #includes}
\usage{
get_stan_code(filename, spaces = 2)
}
\arguments{
\item{filename}{name of the main \code{.stan} file}

\item{spaces}{number of spaces to use when indenting code (default = 2)}
}
\value{
the full program code as a string, formatted using
\code{format_code}
}
\description{
Get full Stan code from a .stan file, possibly containing #includes
}
\seealso{
Other code editing functions: 
\code{\link{format_code}()},
\code{\link{simplify_code}()}
}
\concept{code editing functions}
