% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{indent_code}
\alias{indent_code}
\title{Indent code based on curly brackets}
\usage{
indent_code(code, spaces)
}
\arguments{
\item{code}{code as a string}

\item{spaces}{number of spaces to add to indenting after each opening
curly bracket (default = 2)}
}
\value{
edited code
}
\description{
Indent code based on curly brackets
}
\seealso{
Other code formatting functions: 
\code{\link{format_code}()},
\code{\link{trim_code}()}
}
\concept{code formatting functions}
