% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{justify_line}
\alias{justify_line}
\title{Param justify a line left}
\usage{
justify_line(line, spaces, trim)
}
\arguments{
\item{line}{code line as a string}

\item{spaces}{number of spaces to add to indenting after each opening
curly bracket (default = 2)}

\item{trim}{if this is true, left trimming is done}
}
\value{
edited code line
}
\description{
Param justify a line left
}
\seealso{
Other code formatting helper functions: 
\code{\link{apply_indent}()},
\code{\link{count_indent_lines}()}
}
\concept{code formatting helper functions}
