% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis.R
\name{list_functions}
\alias{list_functions}
\title{Get names of all user-defined functions}
\usage{
list_functions(code, ...)
}
\arguments{
\item{code}{Stan code to be formatted. Ignored if \code{file} is not NULL.}

\item{...}{additional arguments to \code{\link{full_stan_code}}}
}
\value{
a data frame
}
\description{
Get names of all user-defined functions
}
\concept{code analysis functions}
