% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis.R
\name{list_output}
\alias{list_output}
\alias{parameters}
\alias{transformed_parameters}
\alias{generated_quantities}
\title{List output variables of a Stan model}
\usage{
parameters(code = "", file = NULL, verbose = FALSE)

transformed_parameters(code = "", file = NULL, verbose = FALSE)

generated_quantities(code = "", file = NULL, verbose = FALSE)
}
\arguments{
\item{code}{Stan code as a string. Ignored if \code{file} is not NULL.}

\item{file}{File that contains Stan code. If this is
NULL, \code{code} is used as input.}

\item{verbose}{Should some informational messages be printed?}
}
\value{
A data frame with two columns: Name and Type.
}
\description{
Works with transformed and prettified MIR code.
}
\section{Functions}{
\itemize{
\item \code{parameters}: Lists all parameters.

\item \code{transformed_parameters}: Lists all transformed parameters.

\item \code{generated_quantities}: Lists all generated quantities.
}}

