% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis.R
\name{list_params}
\alias{list_params}
\title{List all parameters of a Stan model}
\usage{
list_params(code = "", file = NULL, args = c("--debug-parse"), verbose = FALSE)
}
\arguments{
\item{code}{Stan code as a string. Ignored if \code{file} is not NULL.}

\item{file}{File that contains Stan code. If this is
NULL, \code{code} is used as input.}

\item{args}{Command line arguments to \code{stanc}. When
\code{args = c("--help")}, the help page with all possible options is
returned.}

\item{verbose}{Should some informational messages be printed?}
}
\value{
program code as a formatted string
}
\description{
List all parameters of a Stan model
}
