% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{model_8schools}
\alias{model_8schools}
\title{A hierarchical model for the eight schools data}
\description{
A hierarchical model for the eight schools data
}
\references{
\itemize{
\item Rubin, D. B. (1981).
Estimation in Parallel randomized experiments.
Journal of Educational Statistics, 6, 377-401
\item Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B.,
Vehtari, A. and Rubin, D. B. (2014).
Bayesian Data Analysis: Third Edition.
}
}
\seealso{
Other Stan data and models: 
\code{\link{data_8schools}}
}
\concept{Stan data and models}
