% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\title{Print a message if verbose mode is on}
\usage{
msg(text, verbose, ...)
}
\arguments{
\item{text}{The message to be printed. Will be appended with a line break.}

\item{verbose}{Should the message be printed?}

\item{...}{additional arguments to \code{cat}}
}
\value{
nothing
}
\description{
Print a message if verbose mode is on
}
