% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{place_includes}
\alias{place_includes}
\title{Place possible #includes into Stan code}
\usage{
place_includes(code, parent_dir = getwd(), spaces = 2, verbose = FALSE)
}
\arguments{
\item{code}{the Stan program code as a string}

\item{parent_dir}{parent directory for the \code{#include}d files}

\item{spaces}{number of spaces to use when indenting code (default = 2)}

\item{verbose}{should some messages be printed?}
}
\value{
the full program code as a string, formatted using
\code{format_code}
}
\description{
Place possible #includes into Stan code
}
\seealso{
Other code formatting functions: 
\code{\link{format_code}()},
\code{\link{indent_code}()},
\code{\link{trim_code}()}
}
\concept{code formatting functions}
