% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_io.R
\name{read_file}
\alias{read_file}
\title{Read a text file and return its contents as a string}
\usage{
read_file(file)
}
\arguments{
\item{file}{file name}
}
\value{
file contents as a string
}
\description{
Read a text file and return its contents as a string
}
\seealso{
Other IO functions: 
\code{\link{write_file}()}
}
\concept{IO functions}
