% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{readLines_info}
\alias{readLines_info}
\title{Read lines of a file and show an informational message if verbose is true}
\usage{
readLines_info(file, verbose)
}
\arguments{
\item{file}{path to file}

\item{verbose}{should the message be printed}
}
\value{
same as \code{readLines(file)}
}
\description{
Read lines of a file and show an informational message if verbose is true
}
