% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_hessian.R
\name{rstan_hessian}
\alias{rstan_hessian}
\title{Compute hessian in unconstrained space}
\usage{
rstan_hessian(fit, udraws)
}
\arguments{
\item{fit}{An rstan stanfit}

\item{udraws}{Place(s) to compute Hessian(s) at}
}
\value{
Hessian(s)
}
\description{
If udraws is a vector, the output is the Hessian computed
at the point in parameter space corresponding to udraws
}
\details{
If udraws is a matrix, each row is interpreted as a place
to compute a Hessian and the output is a list of Hessians
}
