% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rstan_fit.R
\name{rstan_make_fit}
\alias{rstan_make_fit}
\title{Build an rstan fit object from a modelfile and a data list}
\usage{
rstan_make_fit(modelfile, data)
}
\arguments{
\item{modelfile}{A stan model file}

\item{data}{Data for stan model}
}
\value{
An rstan fit object
}
\description{
Build an rstan fit object from a modelfile and a data list
}
