% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_power_method.R
\name{rstan_power_method}
\alias{rstan_power_method}
\title{Compute largest magnitude eigenvalue of Hessian and
corresponding eigenvector}
\usage{
rstan_power_method(fit, udraws, L = NULL, max_iterations = 200, tol = 1e-05)
}
\arguments{
\item{fit}{An rstan stanfit}

\item{udraws}{Place(s) to compute Hessian(s) at}
}
\value{
eigenvalue/eigenvector pairs
}
\description{
\itemize{
\item If \code{udraws} is a vector, the output is the eigenvalue/
eigenvector pair computed at the point in parameter
space corresponding to \code{udraws}

\item If \code{udraws} is a matrix, each row is interpreted as a place
to compute an eigenvalue/eigenvector pair and the output
is a list of eigenvalue/eigenvector pairs

\item \code{L} is an optional preconditioner matrix. If this is supplied
compute the eigenvalue/eigenvector pairs of \code{L^T * H * L}

\item \code{max_iterations} is the maximum number of power method iterations
to use on any output

\item \code{tol} is the relative tolerance check done on the eigenvalue
}
}
