% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_simplification.R
\name{simplify_code}
\alias{simplify_code}
\title{Simplify a Stan program}
\usage{
simplify_code(code, data)
}
\arguments{
\item{code}{the Stan code for the program}

\item{data}{the supplied data list (in \code{rstan} format)}
}
\value{
updated code as a string
}
\description{
Simplify a Stan program
}
\concept{code editing functions}
