% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{simplify_code_helpers}
\alias{simplify_code_helpers}
\alias{remove_unused_functions}
\alias{remove_empty_forloops}
\alias{remove_size0_data}
\title{Helper functions for Stan code simplification}
\usage{
remove_unused_functions(code, data)

remove_empty_forloops(code, data)

remove_size0_data(code, data)
}
\arguments{
\item{code}{the Stan code for the program}

\item{data}{the supplied data list (in \code{rstan} format)}
}
\value{
updated code as a string
}
\description{
\itemize{
\item \code{remove_unused_functions} removes user-defined functions that
are not needed
\item \code{remove_empty_forloops} removes for loops that would loop from
1 to 0
\item \code{remove_size0_data} removes data declarations which would have
a dimension of size 0
}
}
