% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{split_code}
\alias{split_code}
\title{Split code at given location}
\usage{
split_code(code, idx = NA)
}
\arguments{
\item{code}{code to split as a string}

\item{idx}{index of the character where to split}
}
\value{
A named list with elements
\itemize{
\item \code{before} - the part before the pattern
\item \code{middle} - the character at the split location
\item \code{after} - the part after the pattern
}
}
\description{
Split code at given location
}
\seealso{
Other code formatting helper functions: 
\code{\link{format_code_helpers}}
}
\concept{code formatting helper functions}
