% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanbreaker.R
\docType{package}
\name{stanbreaker-package}
\alias{stanbreaker-package}
\alias{stanbreaker}
\title{The 'stanbreaker' package.}
\description{
Write description.
}
\references{
\enumerate{
\item Carpenter, B. et al. (2017).
\emph{Stan: A probabilistic programming language}. Journal of Statistical
Software 76(1).
}
}
\author{
Juho Timonen (first.last at iki.fi) and Ben Bales
}
\keyword{Carlo,}
\keyword{Hamiltonian}
\keyword{Monte}
\keyword{Stan,}
\keyword{adaptation,}
\keyword{development}
\keyword{tools}
