% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanc.R
\name{stanc3}
\alias{stanc3}
\title{Call stanc3 using processx}
\usage{
stanc3(code = "", stanc_path = NULL, args = c("--help"), print_stdout = TRUE)
}
\arguments{
\item{code}{Stan code as a character vector}

\item{stanc_path}{Path to the parent directory of \code{stanc}.
If this is \code{NULL}, an attempt is made to find \code{stanc} from the
CmdStan path set by \code{cmdstanr}.}

\item{args}{Command line arguments to \code{stanc}. When
\code{args = c("--help")}, the help page with all possible options is
returned.}

\item{print_stdout}{Should the output of \code{stanc} be printed?}
}
\value{
The output of \code{processx::run} for the \code{stanc} command.
The output value is returned invisibly.
}
\description{
Calls stanc3 using processx. Thanks to Rok Cesnovar.
}
