% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{string_utils}
\alias{string_utils}
\alias{apply_lines}
\alias{ensure_trailing_linebreak}
\title{Common string utilities}
\usage{
apply_lines(txt, fun)

ensure_trailing_linebreak(txt)
}
\arguments{
\item{txt}{a character vector}

\item{fun}{a function that takes one code line (without line break)
as its only argument}
}
\value{
updated character vector
}
\description{
Common string utilities
}
\section{Functions}{
\itemize{
\item \code{apply_lines}: Apply a function to each line of text

\item \code{ensure_trailing_linebreak}: Ensure that a character vector ends in a line break
}}

