% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{trim_code}
\alias{trim_code}
\title{Remove trailing white space from each code line}
\usage{
trim_code(code)
}
\arguments{
\item{code}{Stan code as a string. Ignored if \code{file} is not NULL.}
}
\value{
trimmed code
}
\description{
Remove trailing white space from each code line
}
\seealso{
Other code formatting functions: 
\code{\link{indent_code}()},
\code{\link{place_includes}()}
}
\concept{code formatting functions}
