% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_io.R
\name{write_file}
\alias{write_file}
\title{Write text to a file, overwriting old content}
\usage{
write_file(file, code)
}
\arguments{
\item{file}{file name}

\item{code}{code as a string}
}
\value{
nothing
}
\description{
Write text to a file, overwriting old content
}
\seealso{
Other IO functions: 
\code{\link{read_file}()}
}
\concept{IO functions}
