% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{write_model}
\alias{write_model}
\title{Write stan model code to file}
\usage{
write_model(stan_code, file = NULL)
}
\arguments{
\item{stan_code}{stan code as a string}

\item{file}{path to file}
}
\value{
nothing
}
\description{
Write stan model code to file
}
