% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis_helper.R
\name{check_split}
\alias{check_split}
\title{Check that code split matches original code}
\usage{
check_split(code, rec)
}
\arguments{
\item{code}{full code}

\item{rec}{split code, as a named list with names \code{before},
\code{middle}, and \code{after}}
}
\value{
returns \code{code} invisibly, or throws an error if
\code{paste0(rec$before, rec$middle, rec$after)}
doesn't match the original code
}
\description{
Check that code split matches original code
}
