% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis.R
\name{find_block}
\alias{find_block}
\title{Find a block from Stan code}
\usage{
find_block(code, block)
}
\arguments{
\item{code}{Stan code as a string}

\item{block}{name of block}
}
\value{
a named list with three parts
}
\description{
Find a block from Stan code
}
