% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{format_code}
\alias{format_code}
\title{Format code}
\usage{
format_code(code, spaces = 2)
}
\arguments{
\item{code}{code as a string}

\item{spaces}{number of spaces to add to indenting after each opening
curly bracket (default = 2)}
}
\value{
updated code as a string
}
\description{
This formatting function
\itemize{
\item indents code recursively based on opening and closing curly brackets
\item removes unnesessacy whitespace and line breaks
\item adds some needed linebreaks when missing
}
}
\seealso{
Other code formatting functions: 
\code{\link{format_code_helpers}}
}
\concept{code formatting functions}
