% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{format_code}
\alias{format_code}
\title{Format Stan code}
\usage{
format_code(
  code = "",
  file = NULL,
  use_stanc = TRUE,
  overwrite_file = FALSE,
  place_includes = FALSE,
  stanc_path = NULL,
  spaces = 2,
  verbose = FALSE
)
}
\arguments{
\item{code}{Stan code to be formatted. Ignored if \code{file} is not NULL.}

\item{file}{File that contains the Stan code to be formatted. If this is
NULL, \code{code} is used as input.}

\item{use_stanc}{Should the code be formatted using \code{stanc3} with the
\code{--auto-format} option? If this is
\itemize{
\item TRUE - \code{stanc} must be installed. \emph{Currently, this option
discards all comments from the code}.
\item FALSE - code is formatted by handling strings in R. This option will
not discard comments but is otherwise less sophisticated (currently only
trims lines and indents based on parentheses and curly brackets).
}}

\item{overwrite_file}{Should the file that was given as input be overwritten
by the formatted code?}

\item{place_includes}{Should \code{#include} statements be replaced by their
respective content? (has no effect if \code{use_stanc} is TRUE).}

\item{stanc_path}{Path to the parent directory of \code{stanc}.
If this is \code{NULL}, an attempt is made to find \code{stanc} from the
CmdStan path set by \code{cmdstanr}.}

\item{spaces}{Number of spaces to use for indenting (has no effect if
\code{use_stanc} is TRUE).}

\item{verbose}{Should some informational messages be printed?}
}
\description{
Format Stan code
}
