% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis_helper.R
\name{full_stan_code}
\alias{full_stan_code}
\title{Get full Stan code with includes placed in}
\usage{
full_stan_code(code = "", file = NULL, ...)
}
\arguments{
\item{code}{Stan code as a string. Ignored if \code{file} is not NULL.}

\item{file}{File that contains Stan code. If this is
NULL, \code{code} is used as input.}

\item{...}{additional arguments to \code{\link{format_code}}}
}
\value{
program code as a formatted string
}
\description{
Get full Stan code with includes placed in
}
