% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis.R
\name{get_stan_code}
\alias{get_stan_code}
\title{Get full Stan code from a .stan file}
\usage{
get_stan_code(filename, spaces = 2, verbose = FALSE)
}
\arguments{
\item{filename}{name of the main \code{.stan} file}

\item{spaces}{number of spaces to use when indenting code (default = 2)}

\item{verbose}{should some messages be printed?}
}
\value{
the full program code as a string, formatted using
\code{format_code}, \code{#include}s placed
}
\description{
Get full Stan code from a .stan file
}
\seealso{
Other code analysis functions: 
\code{\link{get_functions}()}
}
\concept{code analysis functions}
