% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{indent_code}
\alias{indent_code}
\alias{indent_part}
\alias{justify_left}
\alias{justify_line}
\alias{locate_closing_bracket}
\title{Indent code recursively based on opening and closing curly brackets}
\usage{
indent_code(code, spaces = 2)

indent_part(code, spaces, indent)

justify_left(code, indent)

justify_line(line, indent)

locate_closing_bracket(code)
}
\arguments{
\item{code}{code as a string}

\item{spaces}{number of spaces to add to indenting after each opening
curly bracket (default = 2)}

\item{indent}{base level of indentation}
}
\value{
updated code as a string
}
\description{
\itemize{
\item \code{indent_code} is the main function
\item \code{indent_part} is a helper function used recursively by
\code{indent_code}
\item \code{justlify_left} is the base case of the recursion, which
justifies the code left at a given number of indentation spaces
\item \code{justify_line} first trims a line by removing leading and
trailing whitespace, and then justifies it correctly
\item \code{locate_closing_bracket} finds the correct closing
bracket
}
}
