% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{indent_code}
\alias{indent_code}
\title{Indent code based on (curly) brackets}
\usage{
indent_code(code, spaces, curly = TRUE)
}
\arguments{
\item{code}{Stan code to be formatted. Ignored if \code{file} is not NULL.}

\item{spaces}{Number of spaces to use for indenting (has no effect if
\code{use_stanc} is TRUE).}

\item{curly}{is indention done based on curly brackets?}
}
\value{
edited code
}
\description{
Indent code based on (curly) brackets
}
\seealso{
Other code formatting functions: 
\code{\link{place_includes}()},
\code{\link{trim_code}()}
}
\concept{code formatting functions}
