% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{place_includes}
\alias{place_includes}
\title{Place possible #includes into Stan code}
\usage{
place_includes(code, include_dir = getwd(), spaces = 2, verbose = FALSE)
}
\arguments{
\item{code}{Stan code to be formatted. Ignored if \code{file} is not NULL.}

\item{include_dir}{Parent directory for possible \code{#include}d files.}

\item{spaces}{Number of spaces to use for indenting (has no effect if
\code{use_stanc} is TRUE).}

\item{verbose}{Should some informational messages be printed?}
}
\value{
the full program code as a string, not formatted or indented
}
\description{
Place possible #includes into Stan code
}
\seealso{
Other code formatting functions: 
\code{\link{indent_code}()},
\code{\link{trim_code}()}
}
\concept{code formatting functions}
