% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{split_code}
\alias{split_code}
\title{Split code at given location or first occurrence of a character}
\usage{
split_code(code, char = "\\n", idx = NULL)
}
\arguments{
\item{code}{code to split as a string}

\item{char}{character where to split (ignored if \code{idx} is not
\code{NULL})}

\item{idx}{index of the character where to split}
}
\value{
A named list with elements
\itemize{
\item \code{before} - the part before the character
\item \code{after} - the part after the character
\item \code{char} - the value of \code{char} given as input, or the
character at location \code{idx}
}
If \code{char} is not found, \code{before} will be the original code
and \code{after} and \code{char} will be an empty strings. If \code{char}
is the first (last) character of \code{code}, then \code{before}
(\code{after}) will be an empty string. The motivation is that
\code{paste0(before, char, after)} will equal the original \code{code}.
}
\description{
Split code at given location or first occurrence of a character
}
\seealso{
Other code formatting functions: 
\code{\link{format_code_helpers}},
\code{\link{format_code}()}
}
\concept{code formatting functions}
