% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{trim_code}
\alias{trim_code}
\title{Remove trailing whitespace from each code line}
\usage{
trim_code(code)
}
\arguments{
\item{code}{code to trim as a string}
}
\value{
trimmed code
}
\description{
Remove trailing whitespace from each code line
}
\seealso{
Other code formatting functions: 
\code{\link{format_code}()},
\code{\link{get_stan_code}()},
\code{\link{indent_code}()}
}
\concept{code formatting functions}
