% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_formatting.R
\name{trim_code}
\alias{trim_code}
\title{Remove trailing whitespace from each code line}
\usage{
trim_code(code)
}
\arguments{
\item{code}{Stan code to be formatted, or name of a file where the
code is.}
}
\value{
trimmed code
}
\description{
Remove trailing whitespace from each code line
}
\seealso{
Other code formatting functions: 
\code{\link{indent_code}()},
\code{\link{place_includes}()}
}
\concept{code formatting functions}
